Templating with concatenation
Is there a way to work with text and merge data with an html file and get an html file with data merge to it as the result? form a programmatic perspective? yes in go this is possible! Remember go encourages using programmatic approach to solve problems. There is a templating approach in go!

These are basic elementary level to doing some templating (merging data with text). You can see html as just text!