Understanding package text/template: parsing & executing templates
Package: Text Template
We want to abstract our code and make it more reusable. Having all our code just inside main function is not the best way to make your code reusable.
https://golang.org/pkg/text/template/

We have a bunch of files, parse them and bring them into our program

Parsing: It means bringing files into your program and using those files.

The way we it is we you parse your templates and store them in a variable of type pointer to a Template. Think of that variable as a container holding all of them templates which you have parsed. An then you can execute those templates