Functions in Templates
We can also pass functions into our templates and we ca use those functions in our templates to do processing on the data that is there in our templates

You can pass in your own functions into a template or there are pre-defined global functions that you can use directly in your template for what you do with the data when it gets to your template.

We use FuncMap to pass our own functions to templates
https://golang.org/pkg/text/template/#FuncMap

When you run go program, think about how th compiler works like is a variable defined before it is used?
How is the compiler reading your code?
As the compiler goes through the code and it parses everything is the function available?