Using methods(with receivers) in Templates
In the pervious programs we have seen we saw how to use pass methods in templates with the help of FuncMap

However we only use FuncMap when the method we want to pass in does not have receivers attached to them

But if the data we pass in is a custom struct type we made and has functions attached to them, we could call those functions in our templates without pass them to func map initially!

NOTE: there are some already global functions in templates you can call but you can pass your own functions through funcMap if they dont have receivers or calling the function right away if they are attached to the type of reciver you passed in

You can have these functions query database before they populate. Whatever you want!