How to Read and Write a Cookie
Cookie is a little file that can hold information which a server can write information to a client machine if the client allows cookies to be written.
Most people allow, but some people do not and thee is a work around for creating state when they do not allow cookies to be written. If the user dont allow cookie, we can pass the unique id through the url with HTTPS so that anything from the client to the server is encrypted. Whether it is cookies or uniqueId passed from url all will be encrypted

A domain can write a cookie, and anytime the browser makes a request back to that domain, it checks if that domain has a cookie and if there is one, it sends the cookie back to that domain as well. But these are domain specific cookies

To set a cookie we call the SetCookie method and pass it a pointer to a cookie
https://golang.org/pkg/net/http/#SetCookie

A cookie is just a struct with fields and has the String() method attached to it that converts the cookie to a string https://golang.org/pkg/net/http/#Cookie

To get a cookie you call the Cookie() method on a request
https://golang.org/pkg/net/http/#Request.Cookie