Creating Sessions
Clients makes request to the server, and we want the client to send a unique identifier (sessionID) to that server. In our server, we could associated that SessionID to a UserID (we have a table that stores a sessionID for every userID). Then we can now take the userID and get anything we want about the user. We could store that Unique ID(sessionID) in a cookie and knowing that cookie are domain specific, whenever the client send a request to our server, if there is a our cookie on their browser, they will send that cookie as well. Then our server can grab that value and we can uniquely identify them.

The SessionID are stored and deleted as the user logs in and out from our server. But the UserID and other information is always there in our DB table