GO & SQL setup
Firstly, be sure to download MySQL server and WorkBench from his link if you don't have one in our system.
We created a new dtatabse(called schema) and added tables to it with few records added to the table all from the MySQL workbench!

Provision your SQL database at AWS.
There are things to note as you Provision the database just for learning and not Prod or dev
- Make sure you select the Free-tier. By selecting this, AWS will disable some other form fields that you dont need to configure because of free tier. Leave the defaults!
- Make your DB instance name, database name and your username thesame. This is just for simplicity and good practise.
- Make your db publicly accessible because we will want to be connecting from our local machine mysql workbench using the assigned url give to us by AWS for this resource. For Prod, you may not want to do this though
- Select the VPC. You can create a new one or use an existing one. Either way, you will need to set the VPC to allow connections from anywhere. On prod or dev, it is good practise to limit the access by listing out the IP address that can only connect to the database
- Be Sure to enable back up, monitoring and also deletion protection

In our code below, we just tested out the connection to the database. And it works fine!